requirements.txt
Flask>=2.0.0
PyYAML>=6.0
python-dotenv>=0.19.0
Werkzeug>=2.0.0
Jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0
